Set up C++ code for querying NNs with VP trees.
