+++ src/init.cpp
@@ -18,6 +18,7 @@ static const R_CallMethodDef all_call_entries[] = {
 
     REGISTER(build_vptree, 1),
     REGISTER(find_vptree, 6),
+    REGISTER(query_vptree, 7),
     {NULL, NULL, 0}
 };
